fix(dashboard): fix after refactor
